Make shared_screen more resilient to refactors
